backend/requirements.txt
alembic>=1.13.0
asyncpg>=0.30.0
chonkie[all]>=1.0.6
fastapi>=0.115.8
fastapi-users[oauth,sqlalchemy]>=14.0.1
firecrawl-py>=1.12.0
github3.py==4.0.1
langchain-community>=0.3.17
langchain-unstructured>=0.1.6
langgraph>=0.3.29
linkup-sdk>=0.2.4
litellm>=1.61.4
markdownify>=0.14.1
notion-client>=2.3.0
pgvector>=0.3.6
playwright>=1.50.0
python-ffmpeg>=2.0.12
rerankers[flashrank]>=0.7.1
sentence-transformers>=3.4.1
slack-sdk>=3.34.0
static-ffmpeg>=2.13
tavily-python>=0.3.2
unstructured-client>=0.30.0
unstructured[all-docs]>=0.16.25
uvicorn[standard]>=0.34.0
validators>=0.34.0
youtube-transcript-api>=1.0.3
pymupdf>=1.25.0
watchdog>=6.0.0
python-magic>=0.4.27
bibtexparser>=1.4.0
scholarly>=1.7.11
crossref-commons>=0.0.7
python-dateutil>=2.6.0
redis>=6.0.0
faiss-cpu>=1.12.0
loguru>=0.7.3
